开发(file): 转辙机检测图片上传接口 /file/upload

+++ service/src/main/java/com/wqtang/file/FileService.java
@@ -8,16 +8,18 @@ import com.wqtang.object.vo.response.file.FileUploadResponse;
 import com.wqtang.util.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.BooleanUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.annotation.Resource;
 import java.io.File;
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -29,8 +31,9 @@ public class FileService {
 
     @Value("${file.rootDirectory}")
     private String rootDirectory;
-    @Value("${file.resultFileName}")
-    private String resultFileName;
+
+    @Resource(name = "restTemplate")
+    private RestTemplate restTemplate;
 
     public ResponseEntity<byte[]> commonDownload(FileCommonDownloadRequest request) throws Exception {
         String fileName = request.getFileName(), filePath = FilenameUtils.concat(rootDirectory, fileName);
@@ -61,30 +64,37 @@ public class FileService {
 
     public FileUploadResponse upload(MultipartFile image) throws Exception {
         // 保存上传的图片文件, 并返回该文件的存储位置的绝对路径
-        String filePath = FileUtils.save(image, rootDirectory);
-        // fixme: 调用python算法并等待解析结果
-        Thread.sleep(6000L);
-        // 解析完成后, 在指定路径处会生成一个result.txt文件
-        File resultFile = new File(FilenameUtils.concat(rootDirectory, resultFileName));
-        while (!resultFile.exists()) {
-            // 循环等待result.txt文件的生成
+        String imagePath = FileUtils.save(image, rootDirectory);
+        // fixme: 调用python算法
+        String serviceUrl = "?imagePath={imagePath}";
+        String resultFileName = restTemplate.exchange(serviceUrl, HttpMethod.GET, new HttpEntity<>(getHttpHeaders()), String.class, imagePath).getBody();
+        if (StringUtils.isBlank(resultFileName)) {
+            throw new BusinessException(ErrorEnum.ERROR, "未检测到所有动静接点，请放平手机重新拍摄");
         }
-        // 读取result.txt文件
+        File resultFile = new File(FilenameUtils.concat(rootDirectory, resultFileName));
+        // 读取算法解析结果文件
         List<String> list = FileUtils.readFileByLine(resultFile);
         if (list.size() != 5) {
             throw new BusinessException(ErrorEnum.ERROR, "未检测到所有动静接点，请放平手机重新拍摄");
         }
-        // 删除result.txt文件
+        // 读取完毕后, 删除该解析结果文件
         FileUtils.delete(resultFile);
         // 封装响应结果信息
         FileUploadResponse response = new FileUploadResponse();
-        response.setFileName(FilenameUtils.getName(filePath));
+        response.setFileName(FilenameUtils.getName(imagePath));
         response.setFlag(list.get(0));
         response.setDepth(list.get(1));
         response.setParamOne(list.get(2));
         response.setParamTwo(list.get(3));
         response.setParamThree(list.get(4));
         return response;
+    }
+
+    private HttpHeaders getHttpHeaders() {
+        HttpHeaders headers = new HttpHeaders();
+        headers.setAccept(Collections.singletonList(MediaType.ALL));
+        headers.setConnection(com.google.common.net.HttpHeaders.KEEP_ALIVE);
+        return headers;
     }
 
 }
